Moved test classes for pieces into correct package
